requirements.txt
torch==2.3.1
torchvision==0.18.1
numpy
Pillow
open3d
huggingface_hub
einops
safetensors
pytorch_metric_learning
pytorch-lightning
termcolor
gradio
viser==0.2.23
tqdm
omegaconf
opencv-python
scipy
requests
trimesh
matplotlib
virtualenv>=20.10.0
lz4
gtsam-develop
ftfy 
regex 